+++ com.unity.cinemachine/Runtime/Components/CinemachineBasicMultiChannelPerlin.cs
@@ -1,5 +1,7 @@
+using System;
 using UnityEngine;
 using UnityEngine.Serialization;
+using Random = UnityEngine.Random;
 
 namespace Cinemachine
 {
@@ -13,10 +15,11 @@ namespace Cinemachine
     /// a large family of different noises using the same profile.
     /// </summary>
     /// <seealso cref="NoiseSettings"/>
-    [DocumentationSorting(DocumentationSortingAttribute.Level.UserRef)]
     [AddComponentMenu("")] // Don't display in add component menu
     [SaveDuringPlay]
-    public class CinemachineBasicMultiChannelPerlin : CinemachineComponentBase
+    [CameraPipeline(CinemachineCore.Stage.Noise)]
+    public class CinemachineBasicMultiChannelPerlin 
+        : CinemachineComponentBase, CinemachineFreeLookModifier.IModifiableNoise
     {
         /// <summary>
         /// Serialized property for referencing a NoiseSettings asset
@@ -46,6 +49,12 @@ namespace Cinemachine
         [Tooltip("Scale factor to apply to the frequencies defined in the NoiseSettings asset.  1 is normal.  "
             + "Larger magnitudes will make the noise shake more rapidly.")]
         public float m_FrequencyGain = 1f;
+
+        (float, float) CinemachineFreeLookModifier.IModifiableNoise.NoiseAmplitudeFrequency 
+        { 
+            get => (m_AmplitudeGain, m_FrequencyGain);
+            set { m_AmplitudeGain = value.Item1; m_FrequencyGain = value.Item2; }
+        }
 
         /// <summary>True if the component is valid, i.e. it has a noise definition and is enabled.</summary>
         public override bool IsValid { get { return enabled && m_NoiseProfile != null; } }