+++ com.unity.cinemachine/Runtime/Core/AxisState.cs
@@ -9,7 +9,6 @@ namespace Cinemachine
     /// Axis state for defining how to react to player input.
     /// The settings here control the responsiveness of the axis to player input.
     /// </summary>
-    [DocumentationSorting(DocumentationSortingAttribute.Level.UserRef)]
     [Serializable]
     public struct AxisState
     {
@@ -57,6 +56,7 @@ namespace Cinemachine
 
         /// <summary>The name of this axis as specified in Unity Input manager.
         /// Setting to an empty string will disable the automatic updating of this axis</summary>
+        [InputAxisNameProperty]
         [FormerlySerializedAs("m_AxisName")]
         [Tooltip("The name of this axis as specified in Unity Input manager. "
             + "Setting to an empty string will disable the automatic updating of this axis")]
@@ -350,7 +350,6 @@ namespace Cinemachine
         public bool HasRecentering { get; set; }
 
         /// <summary>Helper for automatic axis recentering</summary>
-        [DocumentationSorting(DocumentationSortingAttribute.Level.UserRef)]
         [Serializable]
         public struct Recentering
         {