+++ com.unity.cinemachine/Runtime/PostProcessing/CinemachineVolumeSettings.cs
@@ -41,7 +41,6 @@ namespace Cinemachine.PostFX
     /// applying them to the current Post-Processing profile, provided that profile has a
     /// DepthOfField effect that is enabled.
     /// </summary>
-    [DocumentationSorting(DocumentationSortingAttribute.Level.UserRef)]
     [ExecuteAlways]
     [AddComponentMenu("")] // Hide in menu
     [SaveDuringPlay]
@@ -201,8 +200,11 @@ namespace Cinemachine.PostFX
                                 if (focusTarget != null)
                                     focusDistance += (state.FinalPosition - focusTarget.position).magnitude;
                             }
-                            dof.focusDistance.value = Mathf.Max(0, focusDistance);
-                            
+                            focusDistance = Mathf.Max(0, focusDistance);
+                            dof.focusDistance.value = focusDistance;
+#if CINEMACHINE_HDRP_7_3_1
+                            state.Lens.FocusDistance = focusDistance;
+#endif
                             profile.isDirty = true;
                         }
                     }